requirements.txt
aiohttp
aiosignal
annotated-types
anyio
attrs
certifi
charset-normalizer
Cython
decorator
deepgram-sdk
distro
dtw-python
edge-tts
elevenlabs
filelock
frozenlist
fsspec
google-generativeai
groq
h11
httpcore
httpx
idna
imageio
imageio-ffmpeg
Jinja2
llvmlite
MarkupSafe
more-itertools
moviepy<2.0.0
mpmath
multidict
networkx
numba
numpy
openai
openai-whisper
pillow
proglog
pydantic
pydantic_core
python-dotenv
regex
requests
scipy
sniffio
sympy
tiktoken
torch
tqdm
typing_extensions
urllib3<2.0.0
whisper-timestamped
yarl
